fix(review): move frontmatter test to dedicated test_sous_chef_content.py
test_sous_chef_content_no_frontmatter belongs in test_sous_chef_content.py
per the tests/cli/CLAUDE.md index convention for sous-chef content invariants.
